backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
httpx
pytest
pytest-asyncio
